+++ src/FileSystem/FileSystem.hpp
@@ -19,12 +19,12 @@ public:
     FileSystem(FileSystem&&) = default;
     FileSystem& operator=(FileSystem&&) = default;
 
-	std::vector<std::string> dir(const std::string& path = "/") const final;
-    void mkdir(const std::string& dirname) const final;
-    void rmdir(const std::string& dirname) const final;
-    void remove(const std::string& filename) const final;
-	bool exists(const std::string& filepath) const final;
-
+	std::vector<std::string> dir(const std::string& path = "/") const override final;
+    void mkdir(const std::string& dirname) const override final;
+    void rmdir(const std::string& dirname) const override final;
+    void remove(const std::string& filename) const override final;
+	bool exists(const std::string& filepath) const override final;
+    void mount(const std::string& newdir, const std::string& mountpoint) const override final;
 private:
     std::shared_ptr<spdlog::logger> m_logger;
 };