+++ src/FileSystem/IFileSystem.hpp
@@ -13,4 +13,5 @@ public:
     virtual void rmdir(const std::string& dirname) const = 0;
     virtual void remove(const std::string& filename) const = 0;
     virtual bool exists(const std::string& filepath) const = 0;
+    virtual void mount(const std::string& newDir, const std::string& mountPoint) const = 0;
 };